:hammer: Refactor add MultiRuleBook
